@@ -1,72 +1,78 @@
 {
   "name": "CloudFlow",
-  "version": "0.5.2",
+  "version": "0.6.0",
   "license": "Apache-2.0",
+  "private": true,
+  "engines": {
+    "node": "^8.0.0",
+    "yarn": "^1.3.2",
+    "npm": "^5.0.0"
+  },
   "scripts": {
     "ng": "ng",
     "start": "ng serve --ssl --host 0.0.0.0 --port 4200 --publicHost http://0.0.0.0:4200 --disableHostCheck --proxy-config proxy.conf.json",
-    "build": "ng build --aot --prod --output-hashing=bundles",
+    "build": "ng build --aot --prod --output-hashing=bundles --no-progress",
     "test": "ng test",
-    "test-ci": "ng test --singleRun",
-    "test-coverage": "ng test --singleRun --code-coverage --reporters=coverage-istanbul",
-    "lint": "ng lint --type-check",
+    "test-ci": "ng test --watch=false --no-progress",
+    "test-coverage": "ng test --watch-false --code-coverage --reporters=coverage-istanbul",
+    "lint": "ng lint CloudFlow --type-check",
     "e2e": "ng e2e",
     "ci": "npm run lint && npm run test-ci && npm run build"
   },
-  "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.10",
-    "@angular/common": "5.2.10",
-    "@angular/compiler": "5.2.10",
-    "@angular/core": "5.2.10",
-    "@angular/forms": "5.2.10",
-    "@angular/http": "5.2.10",
-    "@angular/platform-browser": "5.2.10",
-    "@angular/platform-browser-dynamic": "5.2.10",
-    "@angular/router": "5.2.10",
-    "@ng-bootstrap/ng-bootstrap": "1.1.2",
-    "angular-oauth2-oidc": "~3.1.0",
+    "@angular/animations": "6.0.7",
+    "@angular/common": "6.0.7",
+    "@angular/compiler": "6.0.7",
+    "@angular/core": "6.0.7",
+    "@angular/forms": "6.0.7",
+    "@angular/http": "6.0.7",
+    "@angular/platform-browser": "6.0.7",
+    "@angular/platform-browser-dynamic": "6.0.7",
+    "@angular/router": "6.0.7",
+    "@ng-bootstrap/ng-bootstrap": "2.2.0",
+    "angular-oauth2-oidc": "~4.0.2",
     "angular-split": "1.0.0-rc.3",
     "awesome-bootstrap-checkbox": "^1.0.1",
-    "bootstrap": "4.1.0",
-    "clipboard": "^1.7.1",
-    "codemirror": "^5.33.0",
-    "core-js": "^2.5.0",
+    "bootstrap": "4.1.1",
+    "codemirror": "^5.39.0",
+    "core-js": "^2.5.7",
     "dagre": "^0.7.4",
     "font-awesome": "^4.7.0",
     "jquery": "^3.2.1",
-    "js-yaml": "^3.9.1",
-    "jsplumb": "2.6.9",
-    "moment": "^2.18.1",
-    "ngx-clipboard": "~10.0.0",
+    "js-yaml": "^3.12.0",
+    "jsplumb": "2.7.9",
+    "moment": "^2.22.2",
+    "ngx-clipboard": "~11.1.1",
     "open-sans-fontface": "^1.4.0",
-    "rxjs": "^5.5.0",
-    "sweetalert2": "^6.11.1",
-    "zone.js": "^0.8.16"
+    "rxjs": "^6.2.1",
+    "rxjs-compat": "^6.2.1",
+    "sweetalert2": "^7.24.4",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "5.2.10",
-    "@angular/language-service": "5.2.10",
-    "@types/codemirror": "^0.0.55",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "6.0.7",
+    "@angular/language-service": "6.0.7",
+    "@types/codemirror": "^0.0.57",
     "@types/dagre": "0.7.39",
     "@types/jasmine": "~2.5.53",
     "@types/jquery": "^3.2.11",
-    "@types/js-yaml": "^3.9.0",
-    "@types/node": "^6.0.60",
-    "codelyzer": "~4.0.1",
+    "@types/js-yaml": "^3.11.1",
+    "@types/node": "^8.0.0",
+    "codelyzer": "~4.4.2",
     "dedent-js": "^1.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
+    "jasmine-core": "~3.1.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.4",
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.8.0",
-    "typescript": "~2.5.3"
+    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^1.2.0",
+    "protractor": "~5.3.2",
+    "ts-node": "~5.0.1",
+    "tslint": "~5.10.0",
+    "typescript": "~2.7.2"
   }
 }
